[ICU-6476] Fixed a bug in DateTimePatternGenerator and updated some test cases. X-Commit-URL: https://ssl.icu-project.org/trac/changeset/24466 Former-commit-id: ad5f5736d74b5e4ac5a75efed9e4d427e48ccf89

+++ icu4j/src/com/ibm/icu/text/DateTimePatternGenerator.java
@@ -1695,7 +1695,7 @@ public class DateTimePatternGenerator implements Freezable, Cloneable {
                 baseOriginal[typeValue] = Utility.repeat(String.valueOf(repeatChar),repeatCount);
                 int subTypeValue = row[2];
                 if (subTypeValue > 0) subTypeValue += field.length();
-                type[typeValue] = (byte) subTypeValue;
+                type[typeValue] = subTypeValue;
             }
             return this;
         }
